+++ dashboard/assets/js/libs/EmptyModule.js
