+++ dashboard/assets/js/libs/plotly-custom.js
